feat: card config on account page
